Phase 3: Redis session manager

+++ backend/app.py
@@ -1,6 +1,11 @@
 from flask import Flask, request, jsonify
 from kubernetes import client, config
+from datetime import datetime, timedelta
 import uuid
+import redis
+import json
+
+r = redis.Redis(host="localhost", port=6379, decode_responses=True)
 
 app = Flask(__name__)
 
@@ -11,9 +16,16 @@ v1 = client.CoreV1Api()
 @app.route("/create-lab", methods=["POST"])
 def create_lab():
     data = request.json
+
     user = data.get("user", "guest")
     lab_type = data.get("lab_type", "python")
 
+    # Preventing multiple sessions per user
+    if r.exists(f"user:{user}:session"):
+        return jsonify({"error":"User already has active lab"}), 400 
+
+    # Session id
+    session_id = str(uuid.uuid4()) 
     pod_name = f"{user}-{lab_type}-{str(uuid.uuid4())[:5]}"
 
     container = client.V1Container(
@@ -37,21 +49,46 @@ def create_lab():
 
     v1.create_namespaced_pod(namespace="default", body=pod)
 
+    # Storing session data
+    session_data = {
+        "user":user,
+        "pod_name":pod_name,
+        "lab_type":lab_type,
+        "start_time":str(datetime.utcnow()),
+        "status": "running"
+    }
+
+    r.set(f"session:{session_id}", json.dumps(session_data))
+    r.set(f"user:{user}:session", session_id)
+    
+    # Setting TTL 15 mins
+    r.expire(f"session:{session_id}", 900)
+
     return jsonify({
         "message": "Lab created",
         "pod_name": pod_name
     })
 
-
 # 🔹 Delete Lab
 @app.route("/delete-lab", methods=["POST"])
 def delete_lab():
     data = request.json
-    pod_name = data.get("pod_name")
+    user = data.get("user")
+    
+    session_id = r.get(f"user:{user}:session")
+
+    if not session_id:
+        return jsonify({"error":"No active session"}), 400
+
+    session_data = json.loads(r.get(f"session:{session_id}"))
+    pod_name = session_data["pod_name"]
 
     v1.delete_namespaced_pod(name=pod_name, namespace="default")
 
-    return jsonify({"message": f"{pod_name} deleted"})
+    r.delete(f"session:{session_id}")
+    r.delete(f"user:{user}:session")
+
+    return jsonify({"message": "Lab deleted"})
 
 
 if __name__ == "__main__":
